backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.30.0,<0.31.0
sqlmodel>=0.0.22,<0.0.23
pydantic-settings>=2.3.0,<2.4.0
httpx>=0.27.0,<0.28.0
psycopg[binary]>=3.1.12,<3.2.0
pyjwt>=2.9.0,<3.0.0
supabase>=2.4.0,<2.5.0
